fix #75 - timeline was leaking through the jButton.  Don't ref the jbutton in the timeline.
Even though there was cleanup in the timeline events if the jButton was not displayable, the fact that the component was not displayable caused trident to not dispatch the needed timeline event to the tracker.  By changing it's 'main object' the timeline engine will broadcast pulses no matter what, so the cleanup  code actually will fire.